+++ x/domain/controllers/domain/domain.go
@@ -39,6 +39,19 @@ func NewController(ctx sdk.Context, k keeper.Keeper, domain string) *Domain {
 	}
 }
 
+// WithConfiguration allows to specify a cached config
+func (c *Domain) WithConfiguration(cfg configuration.Config) *Domain {
+	c.conf = &cfg
+	return c
+}
+
+// WithDomain creates a domain controller with a cached domain
+func (c *Domain) WithDomain(dom types.Domain) *Domain {
+	c.domain = &dom
+	c.domainName = dom.Name
+	return c
+}
+
 // ---------------------- VALIDATION -----------------------------
 
 // Validate validates a domain based on the provided checks
@@ -94,17 +107,17 @@ func (c *Domain) gracePeriodFinished() error {
 	if c.ctx.BlockTime().After(expireTime.Add(gracePeriod)) {
 		return nil
 	}
-	return sdkerrors.Wrapf(types.ErrGracePeriodNotFinished, "domain %s grace period has not finished", c.domain.Name)
+	return sdkerrors.Wrapf(types.ErrDomainGracePeriodNotFinished, "domain %s grace period has not finished", c.domain.Name)
 }
 
-func Owner(addr sdk.AccAddress) ControllerFunc {
+func Admin(addr sdk.AccAddress) ControllerFunc {
 	return func(controller *Domain) error {
-		return controller.ownedBy(addr)
+		return controller.isAdmin(addr)
 	}
 }
 
-// ownedBy makes sure the domain is owned by the provided address
-func (c *Domain) ownedBy(addr sdk.AccAddress) error {
+// isAdmin makes sure the domain is owned by the provided address
+func (c *Domain) isAdmin(addr sdk.AccAddress) error {
 	// assert domain exists
 	if err := c.requireDomain(); err != nil {
 		panic("validation check is not allowed on a non existing domain")
@@ -230,7 +243,7 @@ func DeletableBy(addr sdk.AccAddress) ControllerFunc {
 // deletableBy is the underlying operation used by DeletableBy controller
 func (c *Domain) deletableBy(addr sdk.AccAddress) error {
 	// check if either domain is owned by provided address or if grace period is finished
-	if err := c.Validate(Owner(addr)); err != nil && !c.Condition(GracePeriodFinished) {
+	if err := c.Validate(Admin(addr)); err != nil && !c.Condition(GracePeriodFinished) {
 		return sdkerrors.Wrap(types.ErrUnauthorized, "unable to delete domain not owned if grace period is not finished")
 	}
 	return nil