+++ x/domain/domain_handlers.go
@@ -77,7 +77,7 @@ func handlerMsgTransferDomain(ctx sdk.Context, k keeper.Keeper, msg *types.MsgTr
 	err := c.Validate(
 		domain.MustExist,
 		domain.Type(types.ClosedDomain),
-		domain.Owner(msg.Owner),
+		domain.Admin(msg.Owner),
 		domain.NotExpired,
 		domain.Transferable(msg.TransferFlag),
 	)