+++ x/domain/keeper/account.go
@@ -2,10 +2,11 @@ package keeper
 
 import (
 	"fmt"
+	"time"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/iov-one/iovns"
 	"github.com/iov-one/iovns/x/domain/types"
-	"time"
 )
 
 // contains all the functions to interact with the account store
@@ -121,6 +122,39 @@ func (k Keeper) TransferAccount(ctx sdk.Context, account types.Account, newOwner
 	}
 }
 
+// TransferAccountWithReset transfers the account to aliceAddr new owner after modifying account contents
+func (k Keeper) TransferAccountWithReset(ctx sdk.Context, account types.Account, newOwner sdk.AccAddress, reset bool) {
+	// unmap account to owner
+	err := k.unmapAccountToOwner(ctx, account)
+	if err != nil {
+		panic(fmt.Errorf("indexing error (%#v): %w", account, err))
+	}
+	// unmap account targets
+	err = k.unmapTargetToAccount(ctx, account, account.Targets...)
+	if err != nil {
+		panic(fmt.Errorf("indexing error: (%#v): %w", account, err))
+	}
+	if reset {
+		// update account
+		account.Certificates = nil
+		account.Targets = nil
+		account.MetadataURI = ""
+	}
+	account.Owner = newOwner // transfer owner
+	// save account
+	k.SetAccount(ctx, account)
+	// map account to new owner
+	err = k.mapAccountToOwner(ctx, account)
+	if err != nil {
+		panic(fmt.Errorf("indexing error: (%#v): %w", account, err))
+	}
+	// map accounts new targets
+	err = k.mapTargetToAccount(ctx, account, account.Targets...)
+	if err != nil {
+		panic(fmt.Errorf("indexing error: (%#v): %w", account, err))
+	}
+}
+
 // AddAccountCertificate adds aliceAddr new certificate to the account
 func (k Keeper) AddAccountCertificate(ctx sdk.Context, account types.Account, newCert []byte) {
 	// if not add it to accounts certs
@@ -138,13 +172,13 @@ func (k Keeper) DeleteAccountCertificate(ctx sdk.Context, account types.Account,
 }
 
 // RenewAccount updates an account expiration time
-func (k Keeper) RenewAccount(ctx sdk.Context, account types.Account, accountRenew time.Duration) {
+func (k Keeper) RenewAccount(ctx sdk.Context, account *types.Account, accountRenew time.Duration) {
 	// update account time
 	account.ValidUntil = iovns.TimeToSeconds(
 		iovns.SecondsToTime(account.ValidUntil).Add(accountRenew),
 	)
 	// update account in kv store
-	k.SetAccount(ctx, account)
+	k.SetAccount(ctx, *account)
 }
 
 // ReplaceAccountTargets updates an account targets