+++ x/domain/types/msgs.go
@@ -132,6 +132,9 @@ func (m *MsgDeleteAccount) ValidateBasic() error {
 	if m.Domain == "" {
 		return errors.Wrap(ErrInvalidDomainName, "empty")
 	}
+	if m.Name == "" {
+		return errors.Wrap(ErrOpEmptyAcc, "empty")
+	}
 	// success
 	return nil
 }
@@ -194,6 +197,8 @@ type MsgRegisterAccount struct {
 	Name string
 	// Owner is the owner of the account
 	Owner sdk.AccAddress
+	// Registerer is the user who registers this account
+	Registerer sdk.AccAddress
 	// Targets are the blockchain addresses of the account
 	Targets []BlockchainAddress
 	// Broker is the account that facilitated the transaction
@@ -218,6 +223,9 @@ func (m *MsgRegisterAccount) ValidateBasic() error {
 	if m.Owner.Empty() {
 		return errors.Wrap(ErrInvalidOwner, "empty")
 	}
+	if m.Registerer.Empty() {
+		return errors.Wrap(ErrInvalidRegisterer, "empty")
+	}
 	return nil
 }
 
@@ -228,7 +236,7 @@ func (m *MsgRegisterAccount) GetSignBytes() []byte {
 
 // GetSigners implements sdk.Msg
 func (m *MsgRegisterAccount) GetSigners() []sdk.AccAddress {
-	return []sdk.AccAddress{m.Owner}
+	return []sdk.AccAddress{m.Registerer}
 }
 
 // MsgRegisterDomain is the request used to register new domains
@@ -406,9 +414,9 @@ func (m *MsgReplaceAccountTargets) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{m.Owner}
 }
 
-// MsgSetAccountMetadata is the function used
+// MsgReplaceAccountMetadata is the function used
 // to set accounts metadata
-type MsgSetAccountMetadata struct {
+type MsgReplaceAccountMetadata struct {
 	// Domain is the domain name of the account
 	Domain string
 	// Name is the name of the account
@@ -421,17 +429,17 @@ type MsgSetAccountMetadata struct {
 }
 
 // Route implements sdk.Msg
-func (m *MsgSetAccountMetadata) Route() string {
+func (m *MsgReplaceAccountMetadata) Route() string {
 	return RouterKey
 }
 
 // Type implements sdk.Msg
-func (m *MsgSetAccountMetadata) Type() string {
+func (m *MsgReplaceAccountMetadata) Type() string {
 	return "set_account_metadata"
 }
 
 // ValidateBasic implements sdk.Msg
-func (m *MsgSetAccountMetadata) ValidateBasic() error {
+func (m *MsgReplaceAccountMetadata) ValidateBasic() error {
 	if m.Domain == "" {
 		return errors.Wrapf(ErrInvalidDomainName, "empty")
 	}
@@ -445,12 +453,12 @@ func (m *MsgSetAccountMetadata) ValidateBasic() error {
 }
 
 // GetSignBytes implements sdk.Msg
-func (m *MsgSetAccountMetadata) GetSignBytes() []byte {
+func (m *MsgReplaceAccountMetadata) GetSignBytes() []byte {
 	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(m))
 }
 
 // GetSigners implements sdk.Msg
-func (m *MsgSetAccountMetadata) GetSigners() []sdk.AccAddress {
+func (m *MsgReplaceAccountMetadata) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{m.Owner}
 }
 
@@ -465,6 +473,8 @@ type MsgTransferAccount struct {
 	Owner sdk.AccAddress
 	// NewOwner is the new owner of the account
 	NewOwner sdk.AccAddress
+	// Reset indicates if the accounts content will be resetted
+	Reset bool
 }
 
 // Route implements sdk.Msg
@@ -481,6 +491,9 @@ func (m *MsgTransferAccount) Type() string {
 func (m *MsgTransferAccount) ValidateBasic() error {
 	if m.Domain == "" {
 		return errors.Wrap(ErrInvalidDomainName, "empty")
+	}
+	if m.Name == "" {
+		return errors.Wrap(ErrOpEmptyAcc, "empty")
 	}
 	if m.Owner == nil {
 		return errors.Wrap(ErrInvalidOwner, "empty")