+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog 
 
 ## HEAD
+- fix account controller max renew exceed
+- add tests to account handlers in domain
 - reconcile domain spec
 - treat handler as orchestrator
 - extend keeper functionality
@@ -14,6 +16,13 @@
 - change hasSuperUser to DomainType
 - Open domain's admin is changed from zero address to normal address
 - Recon configuration
+- Recon register account handler
+- Recon transfer account handler
+- Recon delete account handler
+- Recon replace account targets handler
+- Recon add account certs handler
+- Recon delete account cert handler
+- Recon replace metadata handler
 
 ## v0.2.5